Fix uint -> usize in lua52-sys

+++ lua52-sys/src/lib.rs
@@ -87,10 +87,10 @@ pub const LUA_HOOKLINE: c_int = 2;
 pub const LUA_HOOKCOUNT: c_int = 3;
 pub const LUA_HOOKTAILRET: c_int = 4;
 
-pub const LUA_MASKCALL: c_int = 1 << LUA_HOOKCALL as uint;
-pub const LUA_MASKRET: c_int = 1 << LUA_HOOKRET as uint;
-pub const LUA_MASKLINE: c_int = 1 << LUA_HOOKLINE as uint;
-pub const LUA_MASKCOUNT: c_int = 1 << LUA_HOOKCOUNT as uint;
+pub const LUA_MASKCALL: c_int = 1 << LUA_HOOKCALL as usize;
+pub const LUA_MASKRET: c_int = 1 << LUA_HOOKRET as usize;
+pub const LUA_MASKLINE: c_int = 1 << LUA_HOOKLINE as usize;
+pub const LUA_MASKCOUNT: c_int = 1 << LUA_HOOKCOUNT as usize;
 
 #[repr(C)]
 #[allow(missing_copy_implementations)]
